Revert "Commit deu certo caraio"
This reverts commit 47f27bb32d9d1d3649da4662270f9d9c56238878.

+++ Hotel.Infra.Data/Repositories/CidadeRepository.cs
@@ -42,7 +42,7 @@ namespace Hotel.Infra.Data.Repositories
       return lstCidades;
     }
 
-    public async Task<List<CidadeEstadoDto>> GetAllCidadadesEstado(int unidadeId)
+    public async Task<List<CidadeEstadoDto>> GetAllCidadadesEstado(long unidadeId)
     {
       Estado estadoAlias = null;
       Cidade cidadeAlias = null;
